webapi/tct-batterystatus-w3c-tests/batterystatus/w3c/battery-plugging-in-chargingTime-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Battery Test: battery-plugging-in-chargingTime</title>
<meta name="flags" content="interact">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  #note {
    background-color: #fef1b5;
    border: solid 1px #cdab2d;
    padding: 5px;
    margin: 15px;
    display: block;
  }
</style>

<h2>Description</h2>
<p>
  This test validates that onchargingtimechange and chargingTime attributes exist and
  are set to correct values, with corresponding events fired,
  when the charger is plugged in.
</p>

<h2>Preconditions</h2>
<ol>
  <li>
    The device is unplugged from the charger before this test is run.
  </li>
  <li>
    The battery must not be full or reach full capacity before the time the test is run.
  </li>
</ol>

<div id="note">
  <ol>
    <li>
      Plug in the charger and wait for the test to complete.
    </li>
    <li>
      The tests may take long time since the definition of how
      often the chargingtimechange event is fired is left to the implementation.
    </li>
  </ol>
</div>

<div id="log"></div>

<script>

(function() {

  setup({ explicit_timeout: true });

  var onchargingtimechange_test = async_test(
      'When the device is plugged in and its charging time is updated, ' +
      'must set the chargingTime attribute\'s value and fire ' +
      'a chargingtimechange event.');

  var batterySuccess = function (battery) {
    var battery_chargingtime = battery.chargingTime;
    battery.onchargingtimechange = onchargingtimechange_test.step_func(function () {
      assert_less_than(battery.chargingTime, battery_chargingtime,
          'The value of the chargingTime attribute must decrease.');
      onchargingtimechange_test.done();
    });
  };

  var batteryFailure = function (error) {
    onchargingtimechange_test.step(function () {
      assert_unreached(error.message);
    });
    onchargingtimechange_test.done();
  };

  navigator.getBattery().then(batterySuccess, batteryFailure);

})();

</script>
